[spirv] TileAndVectorizePass -> TileAndDistributePass + VectorizePass (#6821)
This commit splits the existing TileAndVectorizePass into two
passes: TileAndDistributePass and VectorizePass. TileAndDistributePass
already exists; so this avoids duplicating the logic in it.

Along the way, refined comments for various passes.
