+++ iree/compiler/Codegen/SPIRV/BUILD
@@ -22,8 +22,8 @@ cc_library(
         "SPIRVLowerExecutableTargetPass.cpp",
         "SPIRVRemoveOneTripTiledLoops.cpp",
         "SPIRVTileAndDistribute.cpp",
-        "SPIRVTileAndVectorize.cpp",
         "SPIRVVectorToCooperativeMatrix.cpp",
+        "SPIRVVectorize.cpp",
         "SPIRVVectorizeLoadStore.cpp",
         "Utils.cpp",
     ],
