+++ iree/compiler/Codegen/SPIRV/SPIRVLowerExecutableTargetPass.cpp
@@ -25,13 +25,16 @@ namespace mlir {
 namespace iree_compiler {
 
 namespace {
-/// Lowers an hal.executable.variant operation to scalar/native-vector
+/// Lowers a hal.executable.variant inner module to SPIR-V scalar/native-vector
 /// code. Invokes different compilation pipeline to
-/// - first lower to scalar/native-vector code
+/// - first lower to scalar/native-vector code,
 /// - then convert to SPIRV dialect.
 class SPIRVLowerExecutableTargetPass
     : public SPIRVLowerExecutableTargetBase<SPIRVLowerExecutableTargetPass> {
  public:
+  SPIRVLowerExecutableTargetPass() = default;
+  SPIRVLowerExecutableTargetPass(const SPIRVLowerExecutableTargetPass &pass) {}
+
   void getDependentDialects(DialectRegistry &registry) const override {
     registry.insert<AffineDialect, gpu::GPUDialect, IREE::HAL::HALDialect,
                     linalg::LinalgDialect, linalg_ext::LinalgExtDialect,
@@ -39,19 +42,14 @@ class SPIRVLowerExecutableTargetPass
                     spirv::SPIRVDialect, vector::VectorDialect>();
   }
 
-  SPIRVLowerExecutableTargetPass() = default;
-  SPIRVLowerExecutableTargetPass(const SPIRVLowerExecutableTargetPass &pass){};
-
   void runOnOperation() override;
 
  private:
   Option<bool> testLoweringConfiguration{
       *this, "test-lowering-configuration",
-      llvm::cl::desc(
-          "Flag used for lit-testing the default configuration set for root "
-          "ops in hal.executable.variants. Defaults to false and is set to "
-          "true "
-          "for lit tests. Not for general usage"),
+      llvm::cl::desc("Flag used for lit-testing the configuration set for root "
+                     "ops in hal.executable.variants. Defaults to false. Set "
+                     "to true for lit tests; not for general usage"),
       llvm::cl::init(false)};
 };
 }  // namespace
@@ -99,13 +97,13 @@ void SPIRVLowerExecutableTargetPass::runOnOperation() {
     OpPassManager &nestedModulePM = executableLoweringPipeline.nest<ModuleOp>();
     switch (*passPipeline) {
       case IREE::HAL::DispatchLoweringPassPipeline::SPIRVDistribute:
-        addSPIRVDistributePassPipeline(nestedModulePM);
+        addSPIRVTileAndDistributePassPipeline(nestedModulePM);
         break;
       case IREE::HAL::DispatchLoweringPassPipeline::SPIRVDistributeToGlobalID:
-        addSPIRVDistributeToGlobalIDPipeline(nestedModulePM);
+        addSPIRVDistributeToGlobalIDPassPipeline(nestedModulePM);
         break;
       case IREE::HAL::DispatchLoweringPassPipeline::SPIRVVectorize:
-        addSPIRVVectorizationPassPipeline(nestedModulePM);
+        addSPIRVTileAndVectorizePassPipeline(nestedModulePM);
         break;
       default:
         llvm_unreachable("Unsupported pipeline on GPU target.");
